requirements.txt
pandas==1.3.5
python-dotenv==1.0.0
yagmail==0.15.293
